pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "openjd-model"
authors = [
  { name = "Amazon Web Services" }
]
dynamic = ["version"]
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
description = "Provides a Python implementation of the data model for Open Job Description's template schemas."
# https://pypi.org/classifiers/
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Operating System :: POSIX :: Linux",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: MacOS",
  "License :: OSI Approved :: Apache Software License",
  "Intended Audience :: Developers",
  "Topic :: Software Development :: Libraries"
]
dependencies = [
  "pyyaml == 6.0.*",
  "pydantic >= 1.10.17",
]

[project.urls]
Homepage = "https://github.com/OpenJobDescription/openjd-model-for-python"
Source = "https://github.com/OpenJobDescription/openjd-model-for-python"

[tool.hatch.build]
artifacts = [
  "*_version.py",
]
only-packages = true

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
version_scheme = "post-release"

[tool.hatch.build.hooks.vcs]
version-file = "_version.py"

[tool.hatch.build.hooks.custom]
path = "hatch_version_hook.py"

[[tool.hatch.build.hooks.custom.copy_map]]
sources = [
  "_version.py",
]
destinations = [
  "src/openjd/model",
]

[tool.hatch.build.targets.sdist]
packages = [
  "src/openjd",
]
only-include = [
  "src/openjd",
]

[tool.hatch.build.targets.wheel]
packages = [
  "src/openjd",
]
only-include = [
  "src/openjd",
]

[tool.mypy]
check_untyped_defs = false
show_error_codes = false
pretty = true
ignore_missing_imports = true
disallow_incomplete_defs = false
disallow_untyped_calls = false
show_error_context = true
strict_equality = false
python_version = 3.9
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = false
# Tell mypy that there's a namespace package at src/openjd
namespace_packages = true
explicit_package_bases = true
mypy_path = "src"

# See: https://docs.pydantic.dev/mypy_plugin/
#  - Helps mypy understand pydantic typing.
plugins = "pydantic.v1.mypy"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
ignore = [
  "E501",
  # Double Check if this should be fixed
  "E731",
]

[tool.ruff.lint.pep8-naming]
classmethod-decorators = [
  "classmethod",
  # pydantic decorators are classmethod decorators
  # suppress N805 errors on classes decorated with them
  "pydantic.validator",
  "pydantic.root_validator",
]

[tool.ruff.lint.isort]
known-first-party = [
  "openjd",
]

[tool.black]
line-length = 100

[tool.pytest.ini_options]
xfail_strict = false
addopts = [
    "-rfEx",
    "--durations=5",
    "--cov=src/openjd/model",
    "--color=yes",
    "--cov-report=html:build/coverage",
    "--cov-report=xml:build/coverage/coverage.xml",
    "--cov-report=term-missing",
    "--numprocesses=auto",
    "--timeout=30"
]


[tool.coverage.run]
branch = true
parallel = true


[tool.coverage.paths]
source = [
  "src/"
]

[tool.coverage.report]
show_missing = true
fail_under = 94

[tool.semantic_release]
# Can be removed or set to true once we are v1
major_on_zero = false
tag_format = "{version}"

[tool.semantic_release.commit_parser_options]
allowed_tags = [
    "build",
    "chore",
    "ci",
    "docs",
    "feat",
    "fix",
    "perf",
    "style",
    "refactor",
    "test",
]
minor_tags = []
patch_tags = [
  "chore",
  "feat",
  "fix",
  "refactor",
]

[tool.semantic_release.publish]
upload_to_vcs_release = false

[tool.semantic_release.changelog]
template_dir = ".semantic_release"

[tool.semantic_release.changelog.environment]
trim_blocks = true
lstrip_blocks = true

[tool.semantic_release.branches.release]
match = "(mainline|release)"
